# api-definition/src/main/resources/static/V1/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Task API
  description: The purpose of this API is for external systems to manage tasks.
servers:
  - url: http://localhost:8080
    description: Task API location
...
paths:
...
  /users:
    get:
      description: |
        These are the users that exist in the API
      tags:
        - Users
      operationId: getUsers
      x-paginated-object:
        - gr.rk.tasks.V1.dto.UserDTO
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
        - in: query
          name: name
          required: false
          schema:
            type: string
        - in: query
          name: email
          required: false
          schema:
            type: string
            format: 'email'
      responses:
        "200":
          description: Paginated users.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                properties:
                  items:
                    $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      description: Create a new User
      tags:
        - Users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "200":
          description: A User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /users/{name}:
    get:
      description: Return specific user
      tags:
        - Users
      operationId: getUser
      parameters:
        - name: name
          in: path
          description: identifier of the user
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      description: Delete the user logical
      tags:
        - "Users"
      operationId: deleteUser
      parameters:
        - name: name
          in: path
          description: identifier of the user
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully logical delete.
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  schemas:
    Page:
      type: object
      properties:
        pageable:
          type: string
          example: INSTANCE
        last:
          type: boolean
        totalPages:
          type: integer
          example: 1
        totalElements:
          type: integer
          example: 5
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        sort:
          type: object
          properties:
            sorted:
              type: boolean
              example: false
            unsorted:
              type: boolean
            empty:
                type: boolean
        first:
          type: boolean
        numberOfElements:
          type: integer
          example: 5
        empty:
          type: boolean
          example: false
...
    Group:
      description: This is a group that a user belongs to
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        realm:
          type: string
          description: This field is the realm from the IDP
          example: Application1
          readOnly: true
        creationDate:
          type: string
          pattern: '^([0-9]+)-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])[Tt]([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$'
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
          readOnly: true
        users:
          readOnly: true
          items:
            $ref: '#/components/schemas/User'
    User:
      description: This is a user object
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        groups:
          readOnly: true
          items:
            $ref: "#/components/schemas/Group"
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum:
            - fail
            - error
          description: |
            The fail status indicating that there was a problem with the data submitted or some
            pre-condition on API call was not satisfied.
            The error status indicating that an error occurred in processing the request, i.e. Exception
            was thrown
        message:
          type: string
    Pageable:
      description: minimal Pageable query parameters
      type: object
      properties:
        page:
          type: integer
          description: Results page you want to retrieve (0..N)
        size:
          type: integer
          description: Number of records per page
        sort:
          description: Sorting criteria in the format -> property,asc|desc no space after comma!.
          example: id,asc
          type: string
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            properties:
              errors:
                items:
                  $ref: '#/components/schemas/Error'